% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{burnt_sienna}
\alias{burnt_sienna}
\title{burnt sienna}
\format{
An object of class \code{character} of length 1.
}
\usage{
burnt_sienna
}
\description{
siei burnt sienna
}
\keyword{datasets}
