% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{burnt_sienna_light}
\alias{burnt_sienna_light}
\title{burnt sienna light}
\format{
An object of class \code{character} of length 1.
}
\usage{
burnt_sienna_light
}
\description{
siei burnt sienna light
}
\keyword{datasets}
