% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cascade}
\alias{cascade}
\title{Cascade Indicators}
\format{
A data frame with six variables: \code{operatingunit},
\code{prime_partner_name}, \code{indicator}, \code{period},
\code{period_type}, and \code{value}.
}
\usage{
cascade
}
\description{
Masked & subset MSD data on the clinical cascade for Saturn
Operating Unit (Total Numerators only). Data sourced from ICPI/TrainingDataset.
}
\keyword{datasets}
