% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.r
\docType{data}
\name{color_caption}
\alias{color_caption}
\title{Color caption default}
\format{An object of class \code{character} of length 1.}
\usage{
color_caption
}
\description{
sets color caption
}
\keyword{datasets}
