% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.r
\docType{data}
\name{color_title}
\alias{color_title}
\title{Title color default}
\format{An object of class \code{character} of length 1.}
\usage{
color_title
}
\description{
sets title color
}
\keyword{datasets}
