% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{font_exists}
\alias{font_exists}
\title{Check if font exists}
\usage{
font_exists(font = c("Source Sans 3", "Source Sans Pro"))
}
\arguments{
\item{font}{typeface to check, defaults to Source Sans 3 and Pro}
}
\description{
Check if font exists
}
\keyword{internal}
