% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{genoa_light}
\alias{genoa_light}
\title{genoa light (green)}
\format{
An object of class \code{character} of length 1.
}
\usage{
genoa_light
}
\description{
siei genoa light
}
\keyword{datasets}
