% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glitr-package.R
\docType{package}
\name{glitr-package}
\alias{glitr}
\alias{glitr-package}
\title{glitr: Adorned SI style plots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Helps create and export ggplot2 charts in the style used by the GH OHA SI team. Includes multiple styles and themes to tweak plots to user needs. Sample testing data also available.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/USAID-OHA-SI/glitr}
  \item \url{https://usaid-oha-si.github.io/glitr/}
  \item Report bugs at \url{https://github.com/USAID-OHA-SI/glitr/issues}
}

}
\author{
\strong{Maintainer}: Tim Essam \email{tessam@usaid.gov}

Authors:
\itemize{
  \item Aaron Chafetz \email{achafetz@usaid.gov}
  \item Baboyma Kagniniwa \email{bkagniniwa@usaid.gov}
}

}
\keyword{internal}
