% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{golden_sand_light}
\alias{golden_sand_light}
\title{golden sand light}
\format{
An object of class \code{character} of length 1.
}
\usage{
golden_sand_light
}
\description{
siei golden sand light
}
\keyword{datasets}
