% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{grey10k}
\alias{grey10k}
\title{grey10k}
\format{
An object of class \code{character} of length 1.
}
\usage{
grey10k
}
\description{
One of the nine shades of grey included. The lower the k-value, the lighter the shade.
grey10k is the lightest grey. hex value: "#E6E7E8".
Useful for light reference lines, boxes or other areas.
}
\seealso{
Other greys: 
\code{\link{grey20k}},
\code{\link{grey30k}},
\code{\link{grey40k}},
\code{\link{grey50k}},
\code{\link{grey60k}},
\code{\link{grey70k}},
\code{\link{grey80k}},
\code{\link{grey90k}}
}
\concept{greys}
\keyword{datasets}
