% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{grey30k}
\alias{grey30k}
\title{grey30k}
\format{
An object of class \code{character} of length 1.
}
\usage{
grey30k
}
\description{
One of the nine shades of grey included. The lower (higher) the k-value, the lighter (darker) the shade.
hex value: "#BCBEC0"
}
\details{
One of the nine shades of grey included.
}
\seealso{
Other greys: 
\code{\link{grey10k}},
\code{\link{grey20k}},
\code{\link{grey40k}},
\code{\link{grey50k}},
\code{\link{grey60k}},
\code{\link{grey70k}},
\code{\link{grey80k}},
\code{\link{grey90k}}
}
\concept{greys}
\keyword{datasets}
