% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{grey90k}
\alias{grey90k}
\title{grey90k}
\format{
An object of class \code{character} of length 1.
}
\usage{
grey90k
}
\description{
One of the nine shades of grey included. The lower (higher) the k-value, the lighter (darker) the shade.
grey90k is the darkest grey.
hex value: "#414042"
}
\seealso{
Other greys: 
\code{\link{grey10k}},
\code{\link{grey20k}},
\code{\link{grey30k}},
\code{\link{grey40k}},
\code{\link{grey50k}},
\code{\link{grey60k}},
\code{\link{grey70k}},
\code{\link{grey80k}}
}
\concept{greys}
\keyword{datasets}
