% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_electric_indigo}
\alias{hw_electric_indigo}
\title{electric_indigo
hemsworth electric_indigo}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_electric_indigo
}
\description{
electric_indigo
hemsworth electric_indigo
}
\seealso{
Other hemsworth: 
\code{\link{hw_hunter}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_midnight_blue}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_slate}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}},
\code{\link{hw_viking}}
}
\concept{hemsworth}
\keyword{datasets}
