% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_hunter}
\alias{hw_hunter}
\title{hunter
hemsworth hunter}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_hunter
}
\description{
hunter
hemsworth hunter
}
\seealso{
Other hemsworth: 
\code{\link{hw_electric_indigo}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_midnight_blue}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}},
\code{\link{hw_viking}}
}
\concept{hemsworth}
\keyword{datasets}
