% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_lavender_haze}
\alias{hw_lavender_haze}
\title{lavender_haze
hemsworth lavender_haze}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_lavender_haze
}
\description{
lavender_haze
hemsworth lavender_haze
}
\seealso{
Other hemsworth: 
\code{\link{hw_electric_indigo}},
\code{\link{hw_hunter}},
\code{\link{hw_midnight_blue}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_slate}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}},
\code{\link{hw_viking}}
}
\concept{hemsworth}
\keyword{datasets}
