% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_midnight_blue}
\alias{hw_midnight_blue}
\title{midnight_blue
hemsworth midnight_blue aka usaid_blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_midnight_blue
}
\description{
midnight_blue
hemsworth midnight_blue aka usaid_blue
}
\seealso{
Other hemsworth: 
\code{\link{hw_electric_indigo}},
\code{\link{hw_hunter}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}},
\code{\link{hw_viking}}
}
\concept{hemsworth}
\keyword{datasets}
