% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_orchid_bloom}
\alias{hw_orchid_bloom}
\title{orchid_bloom
hemsworth orchid_bloom}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_orchid_bloom
}
\description{
orchid_bloom
hemsworth orchid_bloom
}
\seealso{
Other hemsworth: 
\code{\link{hw_electric_indigo}},
\code{\link{hw_hunter}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_midnight_blue}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}},
\code{\link{hw_viking}}
}
\concept{hemsworth}
\keyword{datasets}
