% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_sun_kissed}
\alias{hw_sun_kissed}
\title{sun_kissed
hemsworth sun_kissed}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_sun_kissed
}
\description{
sun_kissed
hemsworth sun_kissed
}
\seealso{
Other hemsworth: 
\code{\link{hw_electric_indigo}},
\code{\link{hw_hunter}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_midnight_blue}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_slate}},
\code{\link{hw_tango}},
\code{\link{hw_viking}}
}
\concept{hemsworth}
\keyword{datasets}
