% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_viking}
\alias{hw_viking}
\title{viking
hemsworth viking aka baldwin scooter_med}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_viking
}
\description{
viking
hemsworth viking aka baldwin scooter_med
}
\seealso{
Other hemsworth: 
\code{\link{hw_electric_indigo}},
\code{\link{hw_hunter}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_midnight_blue}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_slate}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}}
}
\concept{hemsworth}
\keyword{datasets}
