% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{moody_blue_light}
\alias{moody_blue_light}
\title{moody blue light}
\format{
An object of class \code{character} of length 1.
}
\usage{
moody_blue_light
}
\description{
siei moody blue light
}
\keyword{datasets}
