% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{old_rose_light}
\alias{old_rose_light}
\title{old rose light}
\format{
An object of class \code{character} of length 1.
}
\usage{
old_rose_light
}
\description{
siei old rose light
}
\keyword{datasets}
