% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{pantone_blue_iris}
\alias{pantone_blue_iris}
\title{blue iris
Pantone Color of the Year 2008}
\format{
An object of class \code{character} of length 1.
}
\usage{
pantone_blue_iris
}
\description{
blue iris
Pantone Color of the Year 2008
}
\seealso{
Other pantone: 
\code{\link{pantone_aqua_sky}},
\code{\link{pantone_blue_turquoise}},
\code{\link{pantone_cerulean}},
\code{\link{pantone_chili_pepper}},
\code{\link{pantone_classic_blue}},
\code{\link{pantone_emerald}},
\code{\link{pantone_fuchsia_rose}},
\code{\link{pantone_greenery}},
\code{\link{pantone_honeysuckle}},
\code{\link{pantone_illuminating}},
\code{\link{pantone_living_coral}},
\code{\link{pantone_marsala}},
\code{\link{pantone_mimosa}},
\code{\link{pantone_radiant_orchid}},
\code{\link{pantone_rose_quartz}},
\code{\link{pantone_sand_dollar}},
\code{\link{pantone_serenity}},
\code{\link{pantone_tangerine_tango}},
\code{\link{pantone_tigerlily}},
\code{\link{pantone_true_red}},
\code{\link{pantone_turquoise}},
\code{\link{pantone_ultimate_gray}},
\code{\link{pantone_ultra_violet}},
\code{\link{pantone_very_peri}},
\code{\link{pantone_viva_magenta}}
}
\concept{pantone}
\keyword{datasets}
