% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.r
\docType{data}
\name{plot_defaults}
\alias{plot_defaults}
\title{Runs globals for plot configuration}
\format{
An object of class \code{list} of length 5.
}
\usage{
plot_defaults
}
\description{
Global configurations arguments generated here.
}
\keyword{datasets}
