% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_helpers.R
\name{print_si_pals}
\alias{print_si_pals}
\title{Print color palettes available in glitr package}
\usage{
print_si_pals()
}
\description{
Print color palettes available for use
}
\details{
Print available palettes
}
\examples{
\dontrun{
#
View different interpolations of palettes
print_si_pals()
}
}
