% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_helpers.R
\name{scale_color_si}
\alias{scale_color_si}
\title{siei color scales for ggplot2}
\usage{
scale_color_si(
  palette = "genoas",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_color_si(
  palette = "genoas",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'si_palettes' list}

\item{discrete}{whether to use a discrete color palette}

\item{alpha}{controls opacity of colors}

\item{reverse}{if true, reverses the order of palette}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
siei color scales for ggplot2
}
\examples{
\dontrun{
#
library(ggplot2)
library(glitr)

Apply palettes to discrete or continuous colors
ggplot(diamonds) +
geom_point(aes(x = carat, y = price, color = cut)) +
scale_color_si("genoas")

ggplot(diamonds) +
geom_point(aes(x = carat, y = price, color = price)) +
scale_color_si("genoas", discrete = FALSE)



}

}
