% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_helpers.R
\name{scale_fill_si}
\alias{scale_fill_si}
\title{#' si_rampr fill scale for ggplot2}
\usage{
scale_fill_si(
  palette = "genoas",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_fill_si(
  palette = "genoas",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'si_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{sets transparency of colors}

\item{reverse}{if true, reverses the order of palette}

\item{...}{additional arguments to pass to scale_fill_gradientn}
}
\description{
#' si_rampr fill scale for ggplot2
}
\examples{
\dontrun{
library(ggplot2)
library(sf)
library(rnaturalearth)
library(glitr)

ggplot(diamonds) +
  geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_si("genoas")

rnaturalearth::ne_countries(continent = "africa", returnclass = "sf") \%>\%
 add_column(runif = runif(nrow(.))) \%>\%
 ggplot() +
 geom_sf(aes(fill = runif), color = "white", size = 0.25) +
 scale_fill_si("old_roses", discrete = FALSE, reverse = T) +
 si_style_void()
 }
}
