% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{scooter_light}
\alias{scooter_light}
\title{scooter light}
\format{
An object of class \code{character} of length 1.
}
\usage{
scooter_light
}
\description{
siei scooter light
}
\keyword{datasets}
