% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{scooter_med}
\alias{scooter_med}
\title{scooter medium}
\format{
An object of class \code{character} of length 1.
}
\usage{
scooter_med
}
\description{
siei scooter medium
}
\keyword{datasets}
