% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_save.R
\name{si_clear_preview}
\alias{si_clear_preview}
\title{Clear out Plot Previews}
\usage{
si_clear_preview()
}
\description{
Cleans out the temporary folder of any temporary plots. The temporary folder
is removed when the session ends, but it may be useful to remove extraneous
previews taking up space and slowing down processing in the session.
}
\examples{
\dontrun{
library(extrafont)
hts \%>\%
  dplyr::filter(period_type == "results",
                indicator == "HTS_TST_POS") \%>\%
  ggplot2::ggplot(aes(period, value)) +
  ggplot2::geom_col() +
  ggplot2::facet_wrap(~forcats::fct_reorder2(modality, period, value)) +
  ggplot2::labs(x = NULL, y = NULL,
                title = "Saturn experienced a drop in Index testing in FY50Q1" \%>\% toupper,
                subtitle = "Saturn, folks!",
                caption = "Source: FY50Q1i MSD") +
  si_style_nolines()

si_preview()

si_clear_preview() }

}
\seealso{
Other saving: 
\code{\link{si_preview}()},
\code{\link{si_save}()}
}
\concept{saving}
