% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\name{si_colors}
\alias{si_colors}
\title{SI Colors}
\usage{
si_colors()
}
\description{
Defines custom colors to use in SI plots
}
\examples{
#Preview colors
  #library(scales)
  #show_col(c(wapo_lblue, wapo_dblue, wapo_lmauve, wapo_dmauve))

# Preview colors as generated palette
  #new_palette <- colorRampPalette(c(si_lorange, si_dorange))(12)
  #show_col(new_palette)

# Apply the new color palette to a plot with the si_style theme
  #ggplot(iris, aes(Sepal.Length, Sepal.Width, fill = Petal.Length)) +
  #geom_point(shape = 21, color = grey50, size = 3) +
  #scale_fill_gradientn(colours = new_palette) + si_style()
}
\keyword{SI}
\keyword{colors}
