% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{si_dorange}
\alias{si_dorange}
\title{si inspired dark orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
si_dorange
}
\description{
dark shade of orange
}
\keyword{datasets}
