% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_legends.R
\name{si_legend_color}
\alias{si_legend_color}
\title{SI recommended legend style for color aesthetics

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Function attempts to create a well
formatted color-based legend to adorn an existing plot.}
\usage{
si_legend_color()
}
\description{
SI recommended legend style for color aesthetics

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Function attempts to create a well
formatted color-based legend to adorn an existing plot.
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_xyline()
}

}
