% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_legends.R
\name{si_legend_fill}
\alias{si_legend_fill}
\title{Creates an SI recommended legend style for filled aesthics}
\usage{
si_legend_fill()
}
\description{
Creates an SI recommended legend style for filled aesthics
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_xyline()
}

}
