% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{si_lorange}
\alias{si_lorange}
\title{si inspired light orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
si_lorange
}
\description{
light orange
}
\keyword{datasets}
