% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_helpers.R
\name{si_pal}
\alias{si_pal}
\title{siei palette with ramped colors}
\usage{
si_pal(palette = "siei", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'si_palettes' list}

\item{alpha}{controls opacity of color}

\item{reverse}{if true, reverses the order of palette}
}
\description{
Helper function to incorporate siei colors into scale fills
}
