% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{si_palettes}
\alias{si_palettes}
\title{SI-theme color palettes}
\format{
An object of class \code{list} of length 40.
}
\usage{
si_palettes
}
\description{
Sets of color palettes that have been collected by the SI team
}
\details{
SI-themed color palettes

A collection of color palettes gathered by the si team
}
\examples{
# List the palettes available
\dontrun{
names(si_palettes)

# Show palettes interpolated
si_pal("carto_div_earth", n = 10) \%>\% scales::show_col(labels = FALSE)
si_pal("carto_div_earth", n = 100) \%>\% scales::show_col(labels = FALSE)
}
}
\keyword{datasets}
