% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\name{si_preview}
\alias{si_preview}
\title{Palette interpolation preview}
\usage{
si_preview(pal_name, n = 10)
}
\arguments{
\item{pal_name}{name of the palette}

\item{n}{number of interpolation points}
}
\description{
Show palette with n values
}
\details{
Chooses palette to use with ggplot2
}
\examples{
\dontrun{
#
View different interpolations of palettes
si_preview(si_palettes$royal_flow)
si_preview(si_palettes$old_rose_flow, n = 20)
}
}
