% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_helpers.R
\name{si_rampr}
\alias{si_rampr}
\title{Palette interpolation}
\usage{
si_rampr(pal_name = "siei", n, alpha = 1, reverse = FALSE)
}
\arguments{
\item{pal_name}{name of the palette}

\item{n}{number of interpolation points}

\item{alpha}{sets transparency of each color}

\item{reverse}{if true, reverses order}
}
\description{
Generate a palette by linear interpolation.
If a discrete palette is selected, colors are recycled to the length of the desired vector.
}
\details{
Chooses palette to use with ggplot2
}
\examples{
\dontrun{
#
View different interpolations of palettes
si_rampr("royals") \%>\%  scales::show_col()
si_rampr("royals", n = 10) \%>\% scales::show_col()
si_rampr("royals", n = 100)
}
}
