% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_save.R
\name{si_save}
\alias{si_save}
\title{Save Plot}
\usage{
si_save(
  filename,
  plot = last_plot(),
  path = NULL,
  scale = 1,
  width = 10,
  height = 5.625,
  dpi = 320,
  ...
)
}
\arguments{
\item{filename}{File name to create on disk, png unless otherwise specified}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{path}{Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.}

\item{scale}{Multiplicative scaling factor.}

\item{width}{plot width in inches, default = 10}

\item{height}{plot height in inches, default = 5.625}

\item{dpi}{plot resolution}

\item{...}{additional arguments to that pass to ggsave}
}
\value{
exports a file with SI dimensions
}
\description{
Wrapper around ggplot2::ggsave with SI standards
}
\examples{
\dontrun{
library(systemfonts)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style()
#default save
si_save("out/sepal.png")
#defaults to png if not specified (.pdf, .svg, .jpeg)
si_save("out/sepal")
#change the width
si_save("out/sepal.jpeg", width = 5)  }
}
\seealso{
Other saving: 
\code{\link{si_clear_preview}()},
\code{\link{si_preview}()}
}
\concept{saving}
