% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_style.R
\name{si_style}
\alias{si_style}
\title{Create Standard SI Plots}
\usage{
si_style(
  font_title,
  font_subtitle,
  font_plot,
  font_caption,
  facet_space = 2,
  text_scale = 1,
  FO = FALSE,
  ...
)
}
\arguments{
\item{font_title}{font family for titles, defaults to Source Sans Pro/Arial}

\item{font_subtitle}{font family for subtitles, defaults to Source Sans Pro/Arial}

\item{font_plot}{font family for the plot space, defaults to Source Sans Pro/Arial}

\item{font_caption}{font family for captions, defaults to Source Sans Pro/Arial}

\item{facet_space}{controls how far apart facets are from each other}

\item{text_scale}{scalar that will grow/shrink all text defined within}

\item{FO}{make everything black (why, who knows)}

\item{...}{pass any parameters from theme that are not already defined within}
}
\description{
This function adorns your ggplot to convert it into a SI style plot. The default font is Source Sans Pro
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style()
  }
}
