% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_map}
\alias{si_style_map}
\title{Create Standard SI Map}
\usage{
si_style_map()
}
\description{
Create Standard SI Map
}
\examples{
\dontrun{

library(extrafont)
library(tidyverse)
library(glitr)
library(gisr)
library(sf)

sfdf <- gisr::get_admin1("Nigeria") \%>\%
 select(name) \%>\%
 mutate(value = runif(nrow(.), 0, 1))

ggplot() +
   geom_sf(data = sfdf,
           aes(fill = value),
           color = grey10k,
           size = .1) +
   scale_fill_si(palette = "genoas",
                 discrete = FALSE,
                 limits = c(0, 1),
                 labels = scales::percent) +
   labs(title = "NIGERIA - \% OF PLHIV BY STATE",
        subtitle = "States from XYZ Region are the most hit by HIV/AIDS",
        caption = base::paste0("Produced by OHA/SIEI/SI, ON ", base::Sys.Date())) +
  si_style_map()
}

}
