% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_nolines}
\alias{si_style_nolines}
\title{SI style plot with blank canvass}
\usage{
si_style_nolines(...)
}
\arguments{
\item{...}{inherits parameters from [si_style()]}
}
\description{
Incorporates the default si_style graph and strips all gridlines. Useful for when you want to focus attention on a few parts of a plot -- such as a bar graph with fewer than four categories.
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_nolines()
}
}
