% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_transparent}
\alias{si_style_transparent}
\title{Create ggplot with transparent background}
\usage{
si_style_transparent(...)
}
\arguments{
\item{...}{inherits parameters from \code{\link[=si_style]{si_style()}}}
}
\description{
Create ggplot with transparent background
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_transparent()
}

}
