% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_void}
\alias{si_style_void}
\title{Create ggplot with no axes or no grid lines}
\usage{
si_style_void(...)
}
\arguments{
\item{...}{inherits parameters from \code{\link[=si_style]{si_style()}}}
}
\description{
Create ggplot with no axes or no grid lines
}
\examples{
\dontrun{
library(systemfonts)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_void()
}
}
