% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_xgrid}
\alias{si_style_xgrid}
\title{SI style plot with only x-axis gridlines}
\usage{
si_style_xgrid(...)
}
\arguments{
\item{...}{inherits parameters from \code{\link[=si_style]{si_style()}}}
}
\description{
Incorporates the default si_style graph and modifies it to include on x-axis gridlines. Useful for when you are highlighting a vertical comparison, such as rotated bar graphs
}
\examples{
\dontrun{
library(systemfonts)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_xgrid()
}
}
