% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_xline}
\alias{si_style_xline}
\title{SI style plot with x-axis line drawn}
\usage{
si_style_xline()
}
\description{
May be used in place of si_style_blank if you would like to anchor the x-axis.
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_xline()
}
}
