% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{siei_blue}
\alias{siei_blue}
\title{SIEI inspired blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
siei_blue
}
\description{
blue color "#2166AC"
}
\keyword{datasets}
