% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{siei_dgrey}
\alias{siei_dgrey}
\title{SIEI inspired dark grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
siei_dgrey
}
\description{
dark grey "#595959"
}
\keyword{datasets}
