% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{siei_dorange}
\alias{siei_dorange}
\title{SIEI inspired dark orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
siei_dorange
}
\description{
dark shade of orange "#D95709"
}
\keyword{datasets}
