% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{siei_lblue}
\alias{siei_lblue}
\title{SIEI inspired light blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
siei_lblue
}
\description{
light blue "#67A9CF"
}
\keyword{datasets}
