% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{siei_lorange}
\alias{siei_lorange}
\title{SIEI inspired light orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
siei_lorange
}
\description{
light orange "#f5b9a1"
}
\keyword{datasets}
