% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{siei_orange}
\alias{siei_orange}
\title{SIEI inspired orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
siei_orange
}
\description{
normal orange "#EF8A62"
}
\keyword{datasets}
