% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{tango_80}
\alias{tango_80}
\title{tango @ 80\% tint}
\format{
An object of class \code{character} of length 1.
}
\usage{
tango_80
}
\description{
One of the five shades of tango. The number is a percent of the
original color, ranging from 100 (100\%) down to 20 (20\%) at 20\% intervals.
The lower the number, the lighter the tint of the original color.
}
\seealso{
Other SIEI tints: 
\code{\link{electric_indigo_100}},
\code{\link{electric_indigo_20}},
\code{\link{electric_indigo_40}},
\code{\link{electric_indigo_60}},
\code{\link{electric_indigo_80}},
\code{\link{hunter_100}},
\code{\link{hunter_20}},
\code{\link{hunter_40}},
\code{\link{hunter_60}},
\code{\link{hunter_80}},
\code{\link{lavender_haze_100}},
\code{\link{lavender_haze_20}},
\code{\link{lavender_haze_40}},
\code{\link{lavender_haze_60}},
\code{\link{lavender_haze_80}},
\code{\link{midnight_blue_100}},
\code{\link{midnight_blue_20}},
\code{\link{midnight_blue_40}},
\code{\link{midnight_blue_60}},
\code{\link{midnight_blue_80}},
\code{\link{orchid_bloom_100}},
\code{\link{orchid_bloom_20}},
\code{\link{orchid_bloom_40}},
\code{\link{orchid_bloom_60}},
\code{\link{orchid_bloom_80}},
\code{\link{slate_100}},
\code{\link{slate_20}},
\code{\link{slate_40}},
\code{\link{slate_60}},
\code{\link{slate_80}},
\code{\link{sun_kissed_100}},
\code{\link{sun_kissed_20}},
\code{\link{sun_kissed_40}},
\code{\link{sun_kissed_60}},
\code{\link{sun_kissed_80}},
\code{\link{tango_100}},
\code{\link{tango_20}},
\code{\link{tango_40}},
\code{\link{tango_60}},
\code{\link{viking_100}},
\code{\link{viking_20}},
\code{\link{viking_40}},
\code{\link{viking_60}},
\code{\link{viking_80}}
}
\concept{SIEI tints}
\keyword{datasets}
