% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{trolley_grey_light}
\alias{trolley_grey_light}
\title{trolley grey light}
\format{
An object of class \code{character} of length 1.
}
\usage{
trolley_grey_light
}
\description{
SIEI trolley grey light
}
\seealso{
Other SIEI colors: 
\code{\link{burnt_sienna_light}},
\code{\link{burnt_sienna}},
\code{\link{denim_light}},
\code{\link{denim}},
\code{\link{genoa_light}},
\code{\link{genoa}},
\code{\link{golden_sand_light}},
\code{\link{golden_sand}},
\code{\link{moody_blue_light}},
\code{\link{moody_blue}},
\code{\link{old_rose_light}},
\code{\link{scooter_light}},
\code{\link{scooter_med}},
\code{\link{scooter}},
\code{\link{trolley_grey}}
}
\concept{SIEI colors}
\keyword{datasets}
