% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{USAID_dgrey}
\alias{USAID_dgrey}
\title{usaid dark grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
USAID_dgrey
}
\description{
usaid dark grey
}
\keyword{datasets}
