% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{USAID_lgrey}
\alias{USAID_lgrey}
\title{usaid light grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
USAID_lgrey
}
\description{
usaid light grey
}
\keyword{datasets}
