% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_lightblue}
\alias{usaid_lightblue}
\title{usaid light blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_lightblue
}
\description{
usaid light blue
}
\keyword{datasets}
