% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_lightgrey}
\alias{usaid_lightgrey}
\title{usaid light grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_lightgrey
}
\description{
usaid light grey
}
\keyword{datasets}
