% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{USAID_ltblue}
\alias{USAID_ltblue}
\title{usaid light blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
USAID_ltblue
}
\description{
usaid light blue
}
\keyword{datasets}
