% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_medblue}
\alias{usaid_medblue}
\title{usaid mdeblue}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_medblue
}
\description{
usaid medblue
}
\keyword{datasets}
