% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_medgrey}
\alias{usaid_medgrey}
\title{usaid med grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_medgrey
}
\description{
usaid med grey
}
\keyword{datasets}
