% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{USAID_mgrey}
\alias{USAID_mgrey}
\title{usaid medium grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
USAID_mgrey
}
\description{
usaid medium grey
}
\keyword{datasets}
