% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_red}
\alias{usaid_red}
\title{USAID red}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_red
}
\description{
USAID red "#ba0c2f"
}
\seealso{
Other USAID colors: 
\code{\link{usaid_black}},
\code{\link{usaid_blue}},
\code{\link{usaid_darkgrey}},
\code{\link{usaid_lightblue}},
\code{\link{usaid_lightgrey}},
\code{\link{usaid_medblue}},
\code{\link{usaid_medgrey}}
}
\concept{USAID colors}
\keyword{datasets}
