% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_dblue}
\alias{wapo_dblue}
\title{wapo inspired dark blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_dblue
}
\description{
wapo inspired dark blue
}
\keyword{datasets}
