% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_dmauve}
\alias{wapo_dmauve}
\title{wapo inspired dark mauve}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_dmauve
}
\description{
wapo inspired dark mauve
}
\keyword{datasets}
