% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_dorange}
\alias{wapo_dorange}
\title{wapo inspired dark orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_dorange
}
\description{
wapo inspired dark orange
}
\keyword{datasets}
