% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_lblue}
\alias{wapo_lblue}
\title{wapo inspired light blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_lblue
}
\description{
wapo inspired light blue
}
\keyword{datasets}
