% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_lgreen}
\alias{wapo_lgreen}
\title{wapo inspired light green}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_lgreen
}
\description{
wapo inspired light green
}
\keyword{datasets}
