% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_lmauve}
\alias{wapo_lmauve}
\title{wapo inspired light mauve}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_lmauve
}
\description{
wapo inspired light mauve
}
\keyword{datasets}
