% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_lorange2}
\alias{wapo_lorange2}
\title{wapo inspired light orange 2}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_lorange2
}
\description{
wapo inspired light orange
}
\keyword{datasets}
