% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_lpurple}
\alias{wapo_lpurple}
\title{wapo inspired light purple}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_lpurple
}
\description{
wapo inspired purple
}
\keyword{datasets}
