% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_pushpop}
\alias{wapo_pushpop}
\title{wapo inspired pushpop orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_pushpop
}
\description{
wapo inspired push-popish orange
}
\keyword{datasets}
