% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_setup.r
\docType{data}
\name{color_plot_text}
\alias{color_plot_text}
\title{Runs globals for plot configuration}
\format{
An object of class \code{character} of length 1.
}
\usage{
color_plot_text
}
\description{
Global configurations arguments generated here.
}
\keyword{datasets}
