% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{hw_midnight_blue}
\alias{hw_midnight_blue}
\title{midnight_blue
hemsworth midnight_blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
hw_midnight_blue
}
\description{
midnight_blue
hemsworth midnight_blue
}
\seealso{
Other SIEI classic colors: 
\code{\link{burnt_sienna}},
\code{\link{burnt_sienna_light}},
\code{\link{denim}},
\code{\link{denim_light}},
\code{\link{electric_indigo}},
\code{\link{genoa}},
\code{\link{genoa_light}},
\code{\link{golden_sand}},
\code{\link{golden_sand_light}},
\code{\link{hunter}},
\code{\link{hw_electric_indigo}},
\code{\link{hw_hunter}},
\code{\link{hw_lavender_haze}},
\code{\link{hw_orchid_bloom}},
\code{\link{hw_slate}},
\code{\link{hw_sun_kissed}},
\code{\link{hw_tango}},
\code{\link{hw_viking}},
\code{\link{lavender_haze}},
\code{\link{light_grey}},
\code{\link{matterhorn}},
\code{\link{midnight_blue}},
\code{\link{moody_blue}},
\code{\link{moody_blue_light}},
\code{\link{nero}},
\code{\link{old_rose_light}},
\code{\link{orchid_bloom}},
\code{\link{scooter}},
\code{\link{scooter_light}},
\code{\link{scooter_med}},
\code{\link{slate}},
\code{\link{sun_kissed}},
\code{\link{suva_grey}},
\code{\link{tango}},
\code{\link{trolley_grey}},
\code{\link{trolley_grey_light}},
\code{\link{viking}},
\code{\link{whisper}}
}
\concept{SIEI classic colors}
\keyword{datasets}
