% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_helpers.R
\name{scale_color_si}
\alias{scale_color_si}
\title{siei color scales for ggplot2}
\usage{
scale_color_si(
  palette = "genoas",
  alpha = 1,
  discrete = FALSE,
  reverse = FALSE,
  ...
)

scale_color_si(
  palette = "genoas",
  alpha = 1,
  discrete = FALSE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'si_palettes' list}

\item{alpha}{sets transparency of each color}

\item{discrete}{whether to use a discrete color palette}

\item{reverse}{if true, reverses the order of palette}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
SIEI colors available through scale_color_si
}
\examples{
\dontrun{
#
library(ggplot2)
library(glitr)

Apply palettes to discrete or continuous colors
ggplot(diamonds) +
geom_point(aes(x = carat, y = price, color = cut)) +
scale_color_si("genoas", discrete = T)

ggplot(diamonds) +
geom_point(aes(x = carat, y = price, color = price)) +
scale_color_si("genoas")}

}
