% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_legends.R
\name{si_legend_color}
\alias{si_legend_color}
\title{Creates an SI recommended legend style for colored aesthics}
\usage{
si_legend_color()
}
\description{
Creates an SI recommended legend style for colored aesthics
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_xyline()
}

}
