% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_legends.R
\name{si_legend_fill}
\alias{si_legend_fill}
\title{SI legend for fill aesthetics}
\usage{
si_legend_fill()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Function attempts to create a well
formatted color-based legend to adorn an existing plot.
}
\details{
This helper function compliments the various \code{\link[=si_style]{si_style()}} themes.
The main action is to move a fill-based legend to the top left
of the plot and stretch the legend. The same action can be done
using the \code{\link[ggplot2:guides]{ggplot2::guides()}} function.
}
\examples{
\dontrun{
library(systemfonts)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, fill = Sepal.Width)) +
geom_point(shape = 21, size = 4) + si_legend_fill() + si_style()
}

}
