% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\name{si_pal}
\alias{si_pal}
\title{Palette interpolation}
\usage{
si_pal(pal_name, alpha = 1, reverse = FALSE, n = 5)
}
\arguments{
\item{pal_name}{name of the palette}

\item{alpha}{controls opacity of color}

\item{reverse}{if true, reverses order}

\item{n}{number of interpolation points}
}
\description{
Gnerate a palette by linear interpolation.
}
\details{
Chooses palette to use with ggplot2
}
\examples{
#
#View different interpolations of palettes
#si_pal("outbreak_or")(5)
#si_pal("outbreak_or")(10)
#si_pal("outbreak_or")(100)

}
