% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{si_palettes}
\alias{si_palettes}
\title{List of SI Color Palettes}
\format{
An object of class \code{list} of length 74.
}
\usage{
si_palettes
}
\description{
Sets of color palettes that have been collected by the SI team. Includes all
colors from the SIEI data visualization guide and a few additional palettes.
Colors are stored as a named list. Use palette name or position to call in
colors.

Color sets are stored as the name of the first color.
Calling \code{si_palettes$viking_d} will return the five color palette where viking is the base.
Calling \code{si_palettes$viking_c} will return an interpolated ramp of 10 colors for viking
Calling \code{si_palettes$viking_t} will return viking and four tints at 80\%, 60\%, 40\%, and 20\%.
See the SIEI data visualization guide for additional details.
}
\examples{
\dontrun{
library(scales)

# List names of palettes available
names(si_palettes)

show_col(si_palettes$siei)
show_col(si_palettes$lavender_haze_d)
show_col(si_palettes$hunter_t)
show_col(si_palettes$sun_kissed_c)

# Show palettes interpolated
si_rampr("carto_div_earth", n = 10) \%>\% show_col(labels = T)
si_rampr("carto_div_earth", n = 100) \%>\% show_col(labels = FALSE)
}
}
\seealso{
\link{si_rampr} for details on creating interpolated palettes of varying lengths.
}
\keyword{datasets}
