% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{si_palettes}
\alias{si_palettes}
\title{List of SI Color Palettes}
\format{
An object of class \code{list} of length 42.
}
\usage{
si_palettes
}
\description{
Sets of color palettes that have been collected by the SI team.
Includes all colors from the SIEI data visualization guide and a few additional palettes.
Colors are stored as a named list. Use palette name or position to call in colors.

Color sets are stored as the name of the first color.
Calling \code{si_palettes$old_rose} will return the five color pairs where old_rose is the base.
Calling \code{si_palettes$old_roses} will return an interpolated ramp of 11 colors for old_rose.
See the SIEI data visualization guide for additional details.
}
\examples{
\dontrun{
# List names of palettes available
names(si_palettes)

scales::show_col(si_palettes$siei)
scales::show_col(si_palettes$old_roses)

# Show palettes interpolated
si_rampr("carto_div_earth", n = 10) \%>\% scales::show_col(labels = T)
si_rampr("carto_div_earth", n = 100) \%>\% scales::show_col(labels = FALSE)
}
}
\seealso{
\link{si_rampr} for details on creating interpolated palettes of varying lengths.
}
\keyword{datasets}
