% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_save.R
\name{si_preview}
\alias{si_preview}
\title{Preview Plot Output}
\usage{
si_preview(scale = 1, width = 10, height = 5.625, dpi = 320, ...)
}
\arguments{
\item{scale}{Multiplicative scaling factor.}

\item{width}{plot width in inches, default = 10}

\item{height}{plot height in inches, default = 5.625}

\item{dpi}{plot resolution}

\item{...}{additional arguments to that pass to ggsave}
}
\value{
saves and previews a png output
}
\description{
Saves and launches a temporary file to see how the sizing and layout of all
aesthetics, e.g. font size, shape size, etc, look when exported from the IDE
preview.
}
\examples{
\dontrun{
library(systemfonts)
hts \%>\%
  dplyr::filter(period_type == "results",
                indicator == "HTS_TST_POS") \%>\%
  ggplot2::ggplot(aes(period, value)) +
  ggplot2::geom_col() +
  ggplot2::facet_wrap(~forcats::fct_reorder2(modality, period, value)) +
  ggplot2::labs(x = NULL, y = NULL,
                title = "Saturn experienced a drop in Index testing in FY50Q1" \%>\% toupper,
                subtitle = "Saturn, folks!",
                caption = "Source: FY50Q1i MSD") +
  si_style_nolines()

si_preview()

si_clear_preview() }

}
\seealso{
Other saving: 
\code{\link{si_clear_preview}()},
\code{\link{si_save}()}
}
\concept{saving}
