% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_style.R
\name{si_style}
\alias{si_style}
\title{Create Standard SI Plots}
\usage{
si_style(
  font_title = "Source Sans Pro",
  font_subtitle = "Source Sans Pro",
  font_plot = "Source Sans Pro",
  font_caption = "Source Sans Pro",
  facet_space = 2,
  text_scale = 1,
  ...
)
}
\arguments{
\item{font_title}{font family for titles}

\item{font_subtitle}{font family for subtitles}

\item{font_plot}{font family for the plot space}

\item{font_caption}{font family for captions}

\item{facet_space}{controls how far apart facets are from each other}

\item{text_scale}{scalar that will grow/shrink all text defined within}

\item{...}{pass any parameters from theme that are not already defined within}
}
\description{
This function adorns your ggplot to convert it into a SI style plot. The default font is Source Sans Pro
}
\examples{
\dontrun{
library(extrafont)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style()
  }
}
