% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_xyline}
\alias{si_style_xyline}
\title{Create Standard SI Plots showing only both x and y axis lines}
\usage{
si_style_xyline(...)
}
\arguments{
\item{...}{inherits parameters from \code{\link[=si_style]{si_style()}}}
}
\description{
Create Standard SI Plots showing only both x and y axis lines
}
\examples{
\dontrun{
library(systemfonts)
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_xyline()
}

}
