% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_themes.R
\name{si_style_ygrid}
\alias{si_style_ygrid}
\title{SI style plot with only y-axis gridlines}
\usage{
si_style_ygrid()
}
\description{
Incorporates the default si_style graph and modifies it to include on y-axis gridlines. Useful for when you are highlighting a horizontal comparison.
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() + si_style_ygrid()

}
