% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{trolley_grey_light}
\alias{trolley_grey_light}
\title{trolley grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
trolley_grey_light
}
\description{
siei trolley grey
}
\keyword{datasets}
