% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_darkgrey}
\alias{usaid_darkgrey}
\title{usaid dark grey}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_darkgrey
}
\description{
usaid dark grey
}
\keyword{datasets}
