% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{usaid_lightblue}
\alias{usaid_lightblue}
\title{USAID light blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
usaid_lightblue
}
\description{
USAID light blue "#a7c6ed"
}
\seealso{
Other USAID colors: 
\code{\link{usaid_black}},
\code{\link{usaid_blue}},
\code{\link{usaid_darkgrey}},
\code{\link{usaid_lightgrey}},
\code{\link{usaid_medblue}},
\code{\link{usaid_medgrey}},
\code{\link{usaid_red}}
}
\concept{USAID colors}
\keyword{datasets}
