% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{USAID_medblue}
\alias{USAID_medblue}
\title{usaid medium blue}
\format{
An object of class \code{character} of length 1.
}
\usage{
USAID_medblue
}
\description{
usaid medium blue
}
\keyword{datasets}
