% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_colors.R
\docType{data}
\name{wapo_dorange}
\alias{wapo_dorange}
\title{wapo inspired dark orange}
\format{
An object of class \code{character} of length 1.
}
\usage{
wapo_dorange
}
\description{
wapo inspired dark orange "#DC7C33"
}
\seealso{
Other Washington Post inspired: 
\code{\link{wapo_dblue}},
\code{\link{wapo_dmauve}},
\code{\link{wapo_lblue}},
\code{\link{wapo_lgreen}},
\code{\link{wapo_lmauve}},
\code{\link{wapo_lorange}},
\code{\link{wapo_lorange2}},
\code{\link{wapo_lpurple}},
\code{\link{wapo_pushpop}}
}
\concept{Washington Post inspired}
\keyword{datasets}
